Add Netlify support (#2)
* Add Netlify config file
  * Uses Netlify provided env vars for baseUrl
* Configured Netlify to build `master` and all PRs

The plan is to still use GitHub pages (johnhollowell.com/cpsc) for the main hosting as it has no bandwidth limits. Netlify will mainly be used for PR previews, but it will also stay up-to-date with the GitHub pages site.
It will be good to show a super-easy way of hosting everything in GitHub, but also good to show other tools in the ecosystem that provides even more functionality.